requirements.txt
Flask==2.2.5
flask-cors==3.0.10
python-nmap==0.7.1